examples/internet-of-agents-incident-network/pyproject.toml
[project]
name = "internet-of-agents-incident-network"
version = "0.2.0"
description = "Multi-org incident response network using ARC for capability-governed tool access"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.52.0",
  "arc-asgi>=0.1.0",
  "arc-fastapi>=0.1.0",
  "arc-sdk-python>=0.1.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "openai-agents>=0.1.0",
  "pynacl>=1.5.0",
  "uvicorn>=0.34.0",
]

[tool.uv]
package = false

[tool.uv.sources]
arc-asgi = { path = "../../sdks/python/arc-asgi" }
arc-fastapi = { path = "../../sdks/python/arc-fastapi" }
arc-sdk-python = { path = "../../sdks/python/arc-sdk-python" }
